Complete app wit modules
