+++ index.html
@@ -60,7 +60,11 @@
       integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe"
       crossorigin="anonymous"
     ></script>
-    <!-- Js Connection -->
-    <script type="module" src="index.js"></script>
+    <!-- CSS Connection -->
+
+    <!-- <script src="js/date.js"></script> -->
+    <!-- <script src="js/script-class.js"></script> -->
+    <!-- <script src="js/spa.js"></script> -->
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
