+++ index.js
@@ -1,18 +1,18 @@
-import addBook from './modules/addBook.js';
-import BookStore from './modules/bookStore.js';
-import contact from './modules/contact.js';
-import booksList from './modules/booksList.js';
-import removeBook from './modules/removeBooks.js';
+import addBook from './modules/AddBook.js';
+import BookStore from './modules/BookStore.js';
+import Contact from './modules/Contact.js';
 import { mainContainer, navLinks } from './modules/DOMVariables.js';
+import BooksList from './modules/BooksList.js';
+import RemoveBook from './modules/RemoveBook.js';
 import { DateTime } from './node_modules/luxon/build/es6/luxon.js';
 
-document.querySelector('time').innerHTML = DateTime.now().toLocaleString(DateTime.DATETIME_MED);
+document.querySelector('time').innerText = DateTime.now().toLocaleString(DateTime.DATETIME_MED);
 
 const bookStore = new BookStore();
 
-booksList(bookStore);
+BooksList(bookStore);
 
-removeBook(bookStore);
+RemoveBook(bookStore);
 
 navLinks.forEach((navLink) => {
   navLink.addEventListener('click', (e) => {
@@ -24,16 +24,16 @@ navLinks.forEach((navLink) => {
 
     switch (e.target.id) {
       case 'list':
-        booksList(bookStore);
+        BooksList(bookStore);
         break;
       case 'add-new':
         addBook(bookStore);
         break;
       case 'contact':
-        contact(bookStore);
+        Contact(bookStore);
         break;
       default:
-        booksList(bookStore);
+        BooksList(bookStore);
         break;
     }
   });
