+++ modules/DOMVariables.js
