--- package.json
@@ -1,33 +1,0 @@
-{
-  "name": "portfolio",
-  "version": "1.0.0",
-  "description": "<a name=\"readme-top\"></a>",
-  "main": "script.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/andrianarivo/portfolio.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/andrianarivo/portfolio/issues"
-  },
-  "homepage": "https://github.com/andrianarivo/portfolio#readme",
-  "devDependencies": {
-    "babel-eslint": "10.x",
-    "eslint": "7.x",
-    "eslint-config-airbnb-base": "14.x",
-    "eslint-plugin-import": "2.x",
-    "hint": "7.x",
-    "stylelint": "13.x",
-    "stylelint-config-standard": "21.x",
-    "stylelint-csstree-validator": "1.x",
-    "stylelint-scss": "3.x"
-  },
-  "dependencies": {
-    "luxon": "^3.3.0"
-  }
-}
